= PnetCDF file header synchronization mechanism =
== Synchronize header data cached in memory ==

1. Changes to attributes, names of variables and dimensions
The following APIs can modify file header while being called in data mode. Note that these APIs can only modify existing attributes, rename variables or dimensions, given that the new attributes and names do not take more file space than the old ones. Otherwise, these APIs are prohibited in data mode.

    ncmpi_rename_dim(),
    ncmpi_rename_var(),
    ncmpi_copy_att(),
    ncmpi_rename_att(), and
    ncmpii_put_att_<type>()

The above semantics implemented in PnetCDF conform with the netCDF for the corresponding APIs. However, since PnetCDF is designed to handle parallel I/O to shared files, it must ensure data consistency for the file header that is cached in the memory of all MPI processes.  In order to do so, PnetCDF requires the above APIs to be called collectively and use the same argument values while in data mode.  PnetCDF does not automatically check the argument consistency, unless the program is in safe mode, i.e. the environment variable PNETCDF_SAFE_MODE is set to 1. Consistency checking requires one or more MPI collective communication operations, a cost some users may want to avoid.  Please note that the above consistency issue is about the file header data that is cached in the memory. For committing the head changes to the file, see below.

2. Changes to the number of records
At the end of collective put APIs that write record variables, PnetCDF synchronizes the number of records across all processes. This memory synchronization is enforced even if the program is not in the safe mode.  On the other hand for independent APIs, the number of records is not automatically synchronized, as memory synchronization requires MPI collective communication calls. Users are encouraged to call ncmpi_sync_numrecs() to explicitly synchronize the number of records. Note this new API is collective, but can be called in the independent data mode. Number of records is stored in the file header. For committing the number of records to the file, please see below.

== Commit header changes to file ==
In PnetCDF, only root process (rank 0) can write/read/modify the file header.  Thus, committing the changes to file header is done through the root. Before writing to file header, PnetCDF checks the consistency of header data cached in memory across MPI processes.  However, PnetCDF does not commit the header changes immediately to the file, unless the file is opened/created in NC_SHARE mode. If NC_SHARE is not set, the changes to the header data will be cached in memory and flushed to the file when calling ncmpi_close().

If you expect your PnetCDF program to be interrupted or exit in between a file open and close (due to returned error codes or system cash) and wish the file contains most up-to-date metadata (e.g. number of records), NC_SHARE mode should be used.




git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1364 c368ceb8-fd28-0410-a457-e46ee1d7cd4a